requirements.txt
click
pydantic
rich
pyyaml
python-dotenv
requests
psutil
pyfiglet
pwinput
datetime